requirements.txt
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
comm==0.2.2
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
importlib-metadata==7.2.1
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
natsort==8.4.0
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
prompt-toolkit==3.0.47
protobuf==4.24.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pygments==2.18.0
pyjwt==2.0.0
pyparsing==3.1.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
stack_data==0.6.2
sty==1.0.4
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.7
openpyxl
wcwidth==0.2.13
zipp==3.19.2
pandas
numpy
gkeepapi